refactor: move audio player components
